Struct mpl3115::MPL3115A2

pub struct MPL3115A2<I2C> { /* fields omitted */ }

MPL3115A2 driver Will start off deactive and in the PressureAlt config requested

Methods

impl<I2C, E> MPL3115A2<I2C> where
I2C: WriteRead<Error = E> + Write<Error = E>,
E: Debug,

Interrupt setting and status

pub fn new(i2c: I2C, pa: PressureAlt) -> Result<Self, Error<E>>

Create a new LIS2DH12 driver from the given I2C peripheral

pub fn destroy(self) -> I2C

Destroy driver instance, return I2C bus instance

pub fn get_device_id(&mut self) -> Result<u8, Error<E>>

Get the WHO_AM_I register

pub fn activate(&mut self) -> Result<(), Error<E>>

Activate Device,

pub fn deactivate(&mut self) -> Result<(), Error<E>>

De-activate Device,

pub fn change_reading_type(&mut self, pa: PressureAlt) -> Result<(), Error<E>>

Change between altitude and pressure

pub fn take_one_pa_reading(&mut self) -> Result<f32, Error<E>>

Get one (blocking) Pressure or Altitude value

pub fn take_one_temp_reading(&mut self) -> Result<f32, Error<E>>

Get one (blocking) Temperature value

pub fn start_reading(&mut self) -> Result<(), Error<E>>

Clear then set the OST bit which causes the sensor to immediately take another reading Needed to sample faster than 1Hz

pub fn check_pa_reading(&mut self) -> Result<bool, Error<E>>

Check the PDR bit for new data

pub fn check_temp_reading(&mut self) -> Result<bool, Error<E>>

Check the TDR bit for new data

pub fn get_pa_reading(&mut self) -> Result<f32, Error<E>>

Get and process the pressure or altitude data

pub fn get_temp_reading(&mut self) -> Result<f32, Error<E>>

Get and process the temperature data

pub fn set_oversample_rate(&mut self, sample_rate: u8) -> Result<(), Error<E>>

Set the number of samples the device makes before saving the data Call with a rate from 0 to 7. Datasheet calls for 128 but you can set it from 1 to 128 samples. The higher the oversample rate the greater the time between data samples.

Example Times: 0 = 8ms 3 = 30ms 7 = 380ms